+++ nuclia/sdk/search.py
@@ -1,3 +1,4 @@
+import backoff
 import json
 import os
 import warnings
@@ -19,6 +20,7 @@ from nucliadb_models.search import (
     SyncAskResponse,
     ChatModel,
 )
+from nucliadb_sdk import exceptions
 from pydantic import ValidationError
 
 from nuclia.data import get_async_auth, get_auth
@@ -71,6 +73,13 @@ class NucliaSearch:
 
     @kb
     @pretty
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def search(
         self,
         *,
@@ -101,6 +110,13 @@ class NucliaSearch:
 
     @kb
     @pretty
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def find(
         self,
         *,
@@ -140,6 +156,13 @@ class NucliaSearch:
         return ndb.ndb.find(req, kbid=ndb.kbid)
 
     @kb
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def ask(
         self,
         *,
@@ -215,6 +238,13 @@ class NucliaSearch:
         return result
 
     @kb
+    @backoff.on_exception(
+        backoff.expo,
+        exceptions.UnknownError,
+        jitter=backoff.random_jitter,
+        max_tries=3,
+        factor=5,
+    )
     def ask_json(
         self,
         schema: Union[str, Dict[str, Any]],